requirements.txt
streamlit==1.34.0  
beautifulsoup4==4.12.2
trafilatura==1.6.0
httpx==0.28.0
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
openai==1.10.0
aiohttp==3.9.1